data_extract/requirement.txt
jsons
psycopg2-binary